requirements.txt
# VisionAssist Requirements - Updated Architecture
# Core ML and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0

# Computer Vision and Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0

# Web Framework and API
flask>=2.3.0
flask-cors>=4.0.0
flask-limiter>=3.5.0  # Rate limiting
requests>=2.31.0
gunicorn>=21.2.0  # Production WSGI server

# Async Support
asyncio-pool>=0.6.0
aiohttp>=3.8.0

# Database and Caching
redis>=4.6.0
sqlalchemy>=2.0.0
alembic>=1.12.0  # Database migrations

# Natural Language Processing
nltk>=3.8.0

# Security and Validation
marshmallow>=3.20.0  # Input validation
bcrypt>=4.0.0  # Password hashing
pyjwt>=2.8.0  # JWT tokens

# Monitoring and Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Utilities and Performance
tqdm>=4.65.0
psutil>=5.9.0
python-dotenv>=1.0.0  # Environment variables

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0  # Type checking

# Production Deployment
docker>=6.1.0
kubernetes>=27.2.0
